Move day view examples into stories
